+++ tencentcloud/tts/v20190823/models.go
@@ -282,6 +282,9 @@ type TextToVoiceRequestParams struct {
 
 	// 是否开启时间戳功能，默认为false。
 	EnableSubtitle *bool `json:"EnableSubtitle,omitempty" name:"EnableSubtitle"`
+
+	// 断句敏感度，默认值为:0，取值范围:[0,1,2]，值越大则敏感度越低，更易断句，此参数建议不要随意调整，可能会影响合成效果。
+	SegmentRate *uint64 `json:"SegmentRate,omitempty" name:"SegmentRate"`
 }
 
 type TextToVoiceRequest struct {
@@ -324,6 +327,9 @@ type TextToVoiceRequest struct {
 
 	// 是否开启时间戳功能，默认为false。
 	EnableSubtitle *bool `json:"EnableSubtitle,omitempty" name:"EnableSubtitle"`
+
+	// 断句敏感度，默认值为:0，取值范围:[0,1,2]，值越大则敏感度越低，更易断句，此参数建议不要随意调整，可能会影响合成效果。
+	SegmentRate *uint64 `json:"SegmentRate,omitempty" name:"SegmentRate"`
 }
 
 func (r *TextToVoiceRequest) ToJsonString() string {
@@ -349,6 +355,7 @@ func (r *TextToVoiceRequest) FromJsonString(s string) error {
 	delete(f, "SampleRate")
 	delete(f, "Codec")
 	delete(f, "EnableSubtitle")
+	delete(f, "SegmentRate")
 	if len(f) > 0 {
 		return tcerr.NewTencentCloudSDKError("ClientError.BuildRequestError", "TextToVoiceRequest has unknown keys!", "")
 	}